Retrieves the dimension of a server-side R object

Gives the dimensions of an R object on the server-side. This function is similar to R function dim.

Usage

ds.dim(
  x = NULL,
  type = "both",
  datasources = NULL,
  classConsistencyCheck = TRUE
)

Arguments

x

a character string providing the name of the input object.

type

a character string that represents the type of analysis to carry out. If type is set to 'combine', 'combined', 'combines' or 'c', the global dimension is returned. If type is set to 'split', 'splits' or 's', the dimension is returned separately for each study. If type is set to 'both' or 'b', both sets of outputs are produced. Default 'both'.

datasources

a list of DSConnection-class objects obtained after login. If the datasources argument is not specified the default set of connections will be used: see datashield.connections_default.

classConsistencyCheck

logical. If TRUE, checks that the input object has the same class across all studies. Default TRUE.

Value

ds.dim retrieves to the client-side the dimension of the object in the form of a vector where the first element indicates the number of rows and the second element indicates the number of columns.

Details

The function returns the dimension of the server-side input object (e.g. array, matrix or data frame) from every single study and the pooled dimension of the object by summing up the individual dimensions returned from each study.

Server function called: dimDS

See also

Author

DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands

Examples