Gets a pooled statistical mean

This is an internal function.

getPooledMean(dtsources, x)

Arguments

dtsources

a list of DSConnection-class objects obtained after login. If the <datasources> the default set of connections will be used: see datashield.connections_default.

x

a character, the name of a numeric vector

Value

a pooled mean

Details

This function is called to avoid calling the client function 'ds.mean' which may stop the process due to some checks not required when computing a mean inside a function.